phonenumber

+++ app.js
@@ -281,9 +281,9 @@ app.post('/b_v_list', (req, res) => {
     var userId = req.body.userid
     var token = req.body.token;
     console.log(`userId: ${userId}`);
-    userService.delete_order(token, () => {
+    userService.delete_order(token, userId, (phone_number, name) => {
 
-        fbService.otn_message(userId, token, function (){
+        fbService.otn_message(userId, name, phone_number,  token, function (){
             res.status(200).json({
                 status: 'succes'
             });
